+++ StationStops/src/StationStops/Helpers.cs
@@ -1,9 +1,43 @@
 ﻿using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 
 namespace StationStops;
 public static class Helpers
 {
+    /// <summary>
+    /// Get the list of stations from a file
+    /// </summary>
+    /// <param name="path">path to the file</param>
+    /// <returns>list of Stations and if the train stops there</returns>
+    public static List<Station> GetStations(string path)
+    {
+        var stations = new List<Station>();
+
+        var fileLines = File.ReadAllLines(path);
+        var lineNumber = 0;
+        foreach (var line in fileLines)
+        {
+            if (string.IsNullOrEmpty(line))
+                continue;
+
+            var lineParts = line.Split(',');
+            if (lineParts.Length != 2)
+            {
+                continue;
+            }
+
+            var stationName = lineParts[0].Trim();
+            var stationStop = lineParts[1].Trim().ToLower();
+            var isStop = stationStop == "true";
+
+            stations.Add(new Station(stationName, isStop, lineNumber));
+            lineNumber++;
+        }
+
+        return stations;
+    }
+
     /// <summary>
     /// Create a list of adjacent station stops starting at the first found stop. 
     /// Bails as soon as the next station is express.
